Move numeric operation args validators
