+++ misc/python/materialize/output_consistency/input_data/operations/number_operations_provider.py
@@ -15,14 +15,14 @@ from materialize.output_consistency.expression.expression_characteristics import
 from materialize.output_consistency.input_data.params.number_operation_param import (
     NumericOperationParam,
 )
+from materialize.output_consistency.input_data.validators.number_args_validator import (
+    MaxMinusNegMaxArgsValidator,
+    ValueGrowsArgsValidator,
+)
 from materialize.output_consistency.operation.operation import (
     DbFunction,
     DbOperation,
     DbOperationOrFunction,
-)
-from materialize.output_consistency.operation.operation_args_validator import (
-    MaxMinusNegMaxArgsValidator,
-    ValueGrowsArgsValidator,
 )
 
 NUMERIC_OPERATION_TYPES: list[DbOperationOrFunction] = []
